requirements.txt
requests>=2.31.0
livekit
livekit-agents
livekit-plugins-openai
python-dotenv
numpy